requirements.txt
slack-sdk>=3.27.0
anthropic>=0.39.0
python-dotenv>=1.0.0
schedule>=1.2.0
